+++ docs/daily/2024-12-20.md
@@ -7,7 +7,7 @@
 - 컬렉션
 
 ## 💻 예제 코드
-- [캡슐화 연습문제](../../src/main/kotlin/day5/Wand.kt)
+- [캡슐화 연습문제](../../src/main/kotlin/day05/Wand.kt)
 - [캡슐화 연습문제](../../src/main/kotlin/day5/Wizard.kt)
 - [캡슐화 연습문제](../../src/test/kotlin/day5/WandTest.kt)
 - [캡슐화 연습문제](../../src/test/kotlin/day5/WizardTest.kt)
