+++ docs/topics/inheritance.md
@@ -66,7 +66,7 @@
 
 
 ## 💻 예제 코드
-- [상속 수업](../../src/main/kotlin/day6/Basic.kt)
+- [상속 수업](../../src/main/kotlin/day06/Basic.kt)
 - [상속 수업](../../src/main/kotlin/day3/Hero.kt)
 - [상속 수업](../../src/main/kotlin/day6/SuperHero.kt)
 - [UML](../../src/main/kotlin/day6/Game.puml)
