+++ src/main/kotlin/day05/CollectionMain.kt
