+++ src/main/kotlin/day06/Basic.kt
