+++ src/main/kotlin/day06/GreatWizard.kt
