+++ src/main/kotlin/day06/PoisonSlime.kt
