+++ src/main/kotlin/day06/SuperHero.kt
