+++ src/main/kotlin/day07/Asset.kt
