+++ src/main/kotlin/day07/Book.kt
