+++ src/main/kotlin/day07/Character.kt
