+++ src/main/kotlin/day07/IntangibleAsset.kt
