+++ src/main/kotlin/day07/TangibleAsset.kt
