+++ src/main/kotlin/day07/Thing.kt
