+++ src/test/kotlin/day06/GreatWizardTest.kt
@@ -1,4 +1,4 @@
-package day6
+package day06
 
 import org.example.day3.Hero
 import org.example.day5.Wand
